+++ Admin/Component/AdminMenu.php
@@ -13,6 +13,7 @@ use OxidEsales\Codeception\Admin\AdminPanel;
 use OxidEsales\Codeception\Admin\CMSPages;
 use OxidEsales\Codeception\Admin\CoreSettings;
 use OxidEsales\Codeception\Admin\Languages;
+use OxidEsales\Codeception\Admin\Manufacturers;
 use OxidEsales\Codeception\Admin\ModulesList;
 use OxidEsales\Codeception\Admin\Newsletter;
 use OxidEsales\Codeception\Admin\Orders;
@@ -41,6 +42,19 @@ trait AdminMenu
         $I->selectEditFrame();
 
         return new CoreSettings($I);
+    }
+
+    public function openManufacturers(): Manufacturers
+    {
+        $I = $this->user;
+
+        $I->selectNavigationFrame();
+        $I->retryClick(Translator::translate('mxmainmenu'));
+        $I->retryClick(Translator::translate('mxmanufacturer'));
+        $I->selectEditFrame();
+        $I->waitForDocumentReadyState();
+
+        return new Manufacturers($I);
     }
 
     /**