+++ Admin/DataObject/Manufacturer.php
@@ -1,0 +1,69 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace OxidEsales\Codeception\Admin\DataObject;
+
+class Manufacturer
+{
+    private string $title = '';
+    private string $shortDescription = '';
+    private int $sortValue = 0;
+    private bool $active = false;
+    private string $icon = '';
+
+    public function getIcon(): string
+    {
+        return $this->icon;
+    }
+
+    public function setIcon(string $icon): void
+    {
+        $this->icon = $icon;
+    }
+
+    public function getTitle(): string
+    {
+        return $this->title;
+    }
+
+    public function setTitle(string $title): void
+    {
+        $this->title = $title;
+    }
+
+    public function getShortDescription(): string
+    {
+        return $this->shortDescription;
+    }
+
+    public function setShortDescription(string $shortDescription): void
+    {
+        $this->shortDescription = $shortDescription;
+    }
+
+    public function getSortValue(): int
+    {
+        return $this->sortValue;
+    }
+
+    public function setSortValue(int $sortValue): void
+    {
+        $this->sortValue = $sortValue;
+    }
+
+    public function isActive(): bool
+    {
+        return $this->active;
+    }
+
+    public function setActive(bool $active): void
+    {
+        $this->active = $active;
+    }
+}